# dsa-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the DSA Backend application.",
        "title": "DSA Backend API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/problem/crud/create": {
            "put": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Create a new lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problem"
                ],
                "summary": "Create a new lecture entry",
                "parameters": [
                    {
                        "description": "Lecture entry details",
                        "name": "lectureEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/problem.LectureEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/create/{lectureid}/{problemid}": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Register a new problem associated with a lecture",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problem"
                ],
                "summary": "Register a new problem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Problem ID",
                        "name": "problemid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Zip file contains problem resources",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problem registered successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/delete/{lectureid}": {
            "delete": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Delete an existing lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problem"
                ],
                "summary": "Delete an existing lecture entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/update/{lectureid}": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Update an existing lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problem"
                ],
                "summary": "Update an existing lecture entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lecture entry details",
                        "name": "lectureEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/problem.LectureEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "User login with user ID and password. Returns a JWT token if successful.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "user"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful. Returns a JWT token.",
                        "schema": {
                            "$ref": "#/definitions/user.userLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request. This error occurs if the user ID or password is missing or incorrect.",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error. This error occurs if there is an issue with the database or password hashing.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "Get current user information from JWT token",
                "tags": [
                    "user"
                ],
                "summary": "Get current user information",
                "responses": {
                    "200": {
                        "description": "Current user information",
                        "schema": {
                            "$ref": "#/definitions/user.userResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "problem.LectureEntryRequest": {
            "type": "object",
            "required": [
                "deadline",
                "id",
                "start_date",
                "title"
            ],
            "properties": {
                "deadline": {
                    "type": "string",
                    "default": "2025-12-01T10:00:00+09:00"
                },
                "id": {
                    "type": "integer",
                    "default": 0
                },
                "start_date": {
                    "type": "string",
                    "default": "2025-10-01T10:00:00+09:00"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "body": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "response.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "user.userLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "DO NOT modify json name, 'access_token' is required in Swagger UI",
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "token_type": {
                    "description": "DO NOT modify json name, 'token_type' is required in Swagger UI",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/user.userResponse"
                }
            }
        },
        "user.userResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/user/login",
            "scopes": {
                "admin": "Grants any rights related to admin user",
                "manager": "Grants any rights related to manager user",
                "me": "Grants any rights related to the current user"
            }
        }
    }
}